+++ src/app/dashboard/layout.tsx
@@ -1,4 +1,4 @@
-import Header from "@/components/layout/Header";
+import Header from "@/components/layout/header";
 
 interface layoutProps {
   children: React.ReactNode;
