+++ src/components/Card/card-info.tsx
