+++ src/components/icons/verification-icon.tsx
