+++ src/components/layout/Header.tsx
@@ -1,4 +1,4 @@
-import ThemeToggle from "./ThemeToggle";
+import ThemeToggle from "./theme-toggle";
 
 const Header = () => {
   return (
