+++ src/components/modules/authentication/ui/sidebar/hooks/use-sidebar.ts
